+++ src/locales/en_US.ts
@@ -426,6 +426,7 @@ const strings = {
   settings_default_screen_title: 'Show Screen on Login',
   settings_default_screen_home: 'Home',
   settings_default_screen_assets: 'Assets',
+  settings_developer_options_force_la: 'Force Light Account Creation',
   swap_preferred_header: 'Preferred Exchange',
   swap_preferred_cheapest: 'Pick best price',
   swap_preferred_dex: 'Prefer Decentralized',